Align currencies source structure
