+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from "./currencies"
-export * from "./decorator"
+export * from "./currencies/getCurrencyFormatting"
+export * from "./currencies/priceDecorator"
 export * from "./search"
